requirements.txt
asgiref==3.8.1
Django==4.2.13
django-filter==24.2
djangorestframework==3.15.1
importlib_metadata==7.1.0
Markdown==3.6
# psycopg2==2.9.9
psycopg2-binary==2.9.9
sqlparse==0.5.0
typing_extensions==4.11.0
tzdata==2024.1
whitenoise==6.6.0
zipp==3.18.1
